% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbr_col_names.R
\name{abbr_col_names}
\alias{abbr_col_names}
\title{Abbreviate HeRAMS Attribute Table Column Names (internal)}
\usage{
abbr_col_names(tib)
}
\arguments{
\item{tib}{\code{tibble}; futur attribute table}
}
\description{
To avoid truncated names and warnings when writing shapefiles
}
