% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/already_processed.R
\name{already_processed}
\alias{already_processed}
\title{Check processed inputs}
\usage{
already_processed(path, alwaysProcess)
}
\arguments{
\item{path}{character; path of a specific input (e.g. "./data/rDEM")}

\item{alwaysProcess}{logical; should the raw input always be processed, even if it has already been 
processed? If FALSE and if the raw input has already been processed the user is 
interactively asked whether they want to process it again or not.}
}
\description{
Internal function that is used to check if an input has already been processed and if so, to ask the user
if they want to process it again.
}
