% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_hf.R
\name{best_hf}
\alias{best_hf}
\title{best_hf}
\usage{
best_hf(
  workDir,
  catchShp,
  popRaster,
  hfShp,
  hfHfColName,
  catchHfColName,
  nTot,
  adminCheck = FALSE,
  npAdmin = NULL,
  adminShp = NULL,
  adminColName = NULL
)
}
\arguments{
\item{workDir}{character; working directory that contains the inputs}

\item{catchShp}{character; name of the catchment shapefile WITHOUT extension}

\item{popRaster}{character; name of the population raster file WITH extension}

\item{hfShp}{character; name of the health facility shapefile WITHOUT extension}

\item{hfHfColName}{character; name of the health facility name column in the health facility shapefile}

\item{catchHfColName}{character; name of the health facility name column in the catchment shapefile}

\item{nTot}{integer; number of health facilities to be selected}

\item{adminCheck}{logical; whether a minimum of health facilities per administrative unit must be selected. FALSE by default}

\item{npAdmin}{integer; minimum number of health facilities to be selected per administrative unit (required if adminCheck = TRUE)}

\item{adminShp}{character; name of the administrative unit shapefile WITHOUT extention (required if adminCheck = TRUE)}

\item{adminColName}{character; name of the administrative unit name column in the administrative unit shapefile (required if adminCheck = TRUE)}
}
\description{
Select the health facilities that offer the best population coverage using their 
catchments and a raster of population. This function also offers the possibility to ensure that N 
facilities are selected for each administrative unit.
}
