% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_downloaded.R
\name{check_downloaded}
\alias{check_downloaded}
\title{Check Downloaded Inputs}
\usage{
check_downloaded(folders)
}
\arguments{
\item{folders}{character; path of a specific input folder}
}
\description{
Internal function that is used to check if an input has already been downloaded and if so, to ask the user
if they want to download it again.
}
