% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check Input}
\usage{
check_input(mainPath, region, type, print = FALSE)
}
\arguments{
\item{mainPath}{character; the parent directory of the country/region name folder}

\item{region}{character; the country/region name}

\item{type}{character; 'raw' or 'processed' depending on whether the required input is
the raw input or the already processed one.}

\item{print}{logical; should the function just print which inputs are available and which are not available. Internally
\code{print} is set FALSE, but when the function is used directly by the user, print should be set to TRUE.}

\item{return}{character vector when \code{print} = FALSE, and NULL when \code{print} = TRUE}
}
\description{
Check whether a raw or processed input exists
}
