% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Check Inputs}
\usage{
check_inputs(mainPath, location, type, onlyPrint = TRUE)
}
\arguments{
\item{mainPath}{character; the parent directory of the location folder}

\item{location}{character; the location folder name}

\item{type}{character; 'raw' or 'processed' depending on whether the required input is
the raw input or the already processed one.}

\item{onlyPrint}{logical; should the function just print which inputs are available and which are not available. Internally
\code{onlyPrint} is set FALSE, but when the function is used directly by the user, print should be set to TRUE.}
}
\value{
character vector when \code{onlyPrint} = FALSE, and NULL when \code{onlyPrint} = TRUE
}
\description{
Check whether a raw or processed input exists
}
\examples{
# Replace workDir with the actual path to your working directory
\dontrun{
mainPath <- "workDir"
initiate_project(mainPath)}

# Replace myLocation with the location name you are working on (workDir subfolder)
\dontrun{
location <- "myLocation"
check_inputs(mainPath, location, type = "raw")}
}
\keyword{internal}
