% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path_length.R
\name{check_path_length}
\alias{check_path_length}
\title{Check path length}
\usage{
check_path_length(path)
}
\arguments{
\item{path}{character}
}
\description{
Check path length to avoid Windows error
}
