% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ts_table.R
\name{check_ts_table}
\alias{check_ts_table}
\title{Check travel scenario table}
\usage{
check_ts_table(xlsi, xlsiFile, vLc)
}
\arguments{
\item{xlsi}{\code{tibble}; travel scenario table}

\item{xlsiFile}{character; travel scenario file path}

\item{vLc}{numeric vector; unique values of the merged landcover raster}
}
\description{
Check the format and content of the travel scenario table (internal function).
}
\keyword{internal}
