% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_input.R
\name{choose_input}
\alias{choose_input}
\title{Select input}
\usage{
choose_input(folders, msg, mostRecent = FALSE)
}
\arguments{
\item{folders}{character; date and time corresponding to each input folder and obtained with the check_exists function}

\item{msg}{character; used to modify the presentation of the different options}

\item{mostRecent}{logical; should the most recent input be selected? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on date and time.}
}
\value{
character; data and time of the input folder creation
}
\description{
Internal function that is used to select an input when multiple ones are available
}
