% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_list.R
\docType{data}
\name{city_list}
\alias{city_list}
\title{Cities}
\format{
Data frame (attribute table)
}
\source{
https://hub.arcgis.com/datasets/schools-BE::world-urban-areas/about
}
\usage{
city_list
}
\description{
Cities
}
\keyword{datasets}
