% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_input.R
\name{copy_input}
\alias{copy_input}
\title{Copy input}
\usage{
copy_input(mainPath, region, input)
}
\arguments{
\item{mainPath}{character; the parent directory of the country/region name folder}

\item{region}{character; the country name}

\item{input}{character; the absolute path of the input. Can be a single path (corresponding to e.g. a health facility table) 
or a vector of different paths (corresponding to e.g. shapefile documents)}
}
\description{
Copy a manually downloaded input to the project corresponding directory before being processed.
}
\examples{
copy_input("C:/Main_directory", "Mali", "C:/Other_directory/Health_facility.xlsx");

input_files <- c("C:/Other_directory/admin.dbf",
                 "C:/Other_directory/admin.prj",
                 "C:/Other_directory/admin.shp",
                 "C:/Other_directory/admin.shx");
copy_input("C:/Main_directory", "Mali", input_files);
}
