% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_list.R
\docType{data}
\name{country_list}
\alias{country_list}
\title{Countries}
\format{
A tibble: 288 × 621
}
\source{
countrycode::codelist (non empty UN names)
}
\usage{
country_list
}
\description{
Countries
}
\keyword{datasets}
