% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_boundaries.R
\name{download_boundaries}
\alias{download_boundaries}
\title{Download Administrative Boundaries}
\usage{
download_boundaries(
  mainPath,
  country,
  adminLevel,
  type = "gbAuthoritative",
  alwaysDownload = FALSE
)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{country}{character; the country folder name}

\item{adminLevel}{integer; administrative level of the boundaries. From 0 to 3.}

\item{type}{character; data type. Can be 'gbOpen' for geoBoundaries data, 'gbHumanitarian' for UN OCHA CODs data, or
'gbAuthoritative' for UN SALB data (default).}

\item{alwaysDownload}{logical; should the administrative boundaries always be downloaded, even if they have already been 
downloaded? If FALSE and if the administrative boundary shapefile has already been downloaded the user is 
interactively asked whether they want to download it again or not.}
}
\description{
Download the administraive boundary shapefile from \emph{geoboundaries} and copy it to its corresponding folder.
}
\details{
The ISO code used to download the shapefile is retrieved by the internal \code{get_param} function. Another internal function
will query the geoBoundaries database (geoboundaries_query) to check availability and retrieve a json file that contains the download link.
}
\examples{
# Replace workDir with the actual path to your working directory
\dontrun{
mainPath <- "workDir"
initiate_project(mainPath, country)}

# Replace myCountry with the country name you are working on (workDir subfolder)
\dontrun{
country <- "myCountry"
download_boundaries(mainPath, country, adminLevel = 1, type = "gbOpen", alwaysDownload = TRUE)}
}
\references{
Runfola, D. et al. (2020) geoBoundaries: A global database of political administrative boundaries. PLoS ONE 15(4): 
e0231866. https://doi.org/10.1371/journal.pone.0231866
}
