% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fictitious_herams_data}
\alias{fictitious_herams_data}
\title{Fictitious example of an HeRAMS health facility data table}
\format{
A tibble with 86 rows and 824 variables:
\describe{
  \item{Health facility type}{Hospital, clinic, infirmary etc.}
  \item{Funcionality status}{Fully functionning, partially functionning, non-functionning}
  ...
}
}
\usage{
fictitious_herams_data
}
\description{
A table that contains health facility information.
}
\keyword{datasets}
