% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fictitious_data_code.R
\docType{data}
\name{fictitious_herams_data_code}
\alias{fictitious_herams_data_code}
\title{Fictitious example of an HeRAMS health facility data table (with codes)}
\format{
A tibble with 86 rows and 824 variables:
\describe{
  \item{MoSD4}{A1, A2, A3}
  \item{HFFUNCT}{A1, A2, A3}
  ...
}
}
\usage{
fictitious_herams_data_code
}
\description{
A table that contains health facility information.
}
\keyword{datasets}
