% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fictitious_data_txt.R
\docType{data}
\name{fictitious_herams_data_txt}
\alias{fictitious_herams_data_txt}
\title{Fictitious example of an HeRAMS health facility data table (with labels)}
\format{
A tibble with 86 rows and 824 variables:
\describe{
  \item{MoSD3}{Hospital, clinic, infirmary etc.}
  \item{HFFUNCT}{Fully functionning, partially functionning, non-functionning}
  ...
}
}
\usage{
fictitious_herams_data_txt
}
\description{
A table that contains health facility information.
}
\keyword{datasets}
