% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fictitious_data_txt_colnames.R
\docType{data}
\name{fictitious_herams_data_txt_colnames}
\alias{fictitious_herams_data_txt_colnames}
\title{Fictitious example of an HeRAMS health facility data table (with full column names)}
\format{
A tibble with 86 rows and 824 variables:
\describe{
  \item{Health facility type}{Hospital, clinic, infirmary etc.}
  \item{Functionality Status of the MoSD}{Fully functionning, partially functionning, non-functionning}
  ...
}
}
\usage{
fictitious_herams_data_txt_colnames
}
\description{
A table that contains health facility information.
}
\keyword{datasets}
