% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_facilities.R
\name{filter_facilities}
\alias{filter_facilities}
\title{Filter facilities}
\usage{
filter_facilities(tib, var, outFolder)
}
\arguments{
\item{tib}{tibble;}

\item{var}{character vector; variable names (from filter_hf function)}

\item{outFolder}{character; output folder}
}
\value{
A tibble
}
\description{
Internal function used in the filter_hf function.
}
