% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hf.R
\name{filter_hf}
\alias{filter_hf}
\title{Filter health facilities}
\usage{
filter_hf(mainPath, region, mostRecent)
}
\arguments{
\item{mainPath}{character; the parent directory of the country/region name folder}

\item{region}{character; the country name}

\item{mostRecent}{logical; should the most recent raw health facility table be used? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on date and time.}
}
\description{
Filters interactively the HeRAMS health facility raw table based on a set of variables and export a table that contains only 
the selected facilities. The selection is recorded within a log.txt file stored in the output folder.
}
