% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoboundaries_query.R
\name{geoboundaries_query}
\alias{geoboundaries_query}
\title{Query geoBoundaries}
\usage{
geoboundaries_query(iso, adminLevel, type, validTypes)
}
\arguments{
\item{iso}{character; country ISO code}

\item{adminLevel}{integer; administrative level of the boundaries. From 0 to 3.}

\item{type}{character; data type. Can be 'gbOpen' for geoBoundaries data, 'gbHumanitarian' for UN OCHA CODs data, or
'gbAuthoritative' for UN SALB data (default).}

\item{validTypes}{character; valid data types (see above)}
}
\description{
Internal function that is used to check availability of boundary shapefile and get the json file
}
\keyword{internal}
