% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Get boundary shapefile}
\usage{
get_boundaries(mainPath, region, type, mostRecent)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{region}{character; the country folder name}

\item{type}{character; 'raw' or 'processed' depending on whether the required input is
the raw input or the already processed one (i.e. projected shapefile).}

\item{mostRecent}{logical; should the most recent input be selected? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on file creation time.}
}
\value{
object of class \code{sf} when a layer was successfully read
}
\description{
Internal function that is used to get the boundary shapefile required in other functions
}
