% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeRAMS_correct_coordinates.R
\name{HeRAMS_correct_coordinates}
\alias{HeRAMS_correct_coordinates}
\title{Correct Coordinates (HeRAMS filtered table)}
\usage{
HeRAMS_correct_coordinates(heramsCSV, corrCSV, gpsX, gpsY, byX, byY)
}
\arguments{
\item{heramsCSV}{character; path to the 'herams_facilities.csv' file (to be modified)}

\item{corrCSV}{character; path to the table that contains correct coordinates}

\item{gpsX}{character; vector of length 2 corresponding to the GPS column names in heramsCSV}

\item{gpsY}{character; vector of length 2 corresponding to the GPS column names in corrCSV}

\item{byX}{character; column name in heramsCSV used for matching rows}

\item{byY}{character; column name in CorrCSV used for matching rows}
}
\description{
Correct or complete coordinates in the herams_facilities.csv based on a second CSV table that contains correct coordinates
}
