% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeRAMS_filter_hf.R
\name{HeRAMS_filter_hf}
\alias{HeRAMS_filter_hf}
\title{Filter health facilities}
\usage{
HeRAMS_filter_hf(
  mainPath,
  country,
  pathTableCode = NULL,
  pathTableText = NULL,
  scenario = NULL,
  mostRecentObs = NULL,
  defaultParameters = TRUE,
  type = TRUE,
  ownership = FALSE,
  status = TRUE,
  building = FALSE,
  equipment = FALSE,
  functionality = FALSE,
  accessibility = FALSE,
  support = FALSE,
  services = FALSE,
  partners = FALSE,
  barriers = FALSE
)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{country}{character; the country folder name}

\item{pathTableCode}{character; path to the HeRAMS CSV table with text for responses ("Available")}

\item{pathTableText}{character; path to the HeRAMS CSV table with codes for responses ("A1")}

\item{scenario}{character; a string of three characters that correspond to the scenario folder suffix like '001', '002'...'010'...'099'...'100'
The criteria for selection are those of the specified scenario. If NULL, an interactive selection by attribute is run in the console.}

\item{mostRecentObs}{logical; should the most recent observation per health facility be taken into account? If NULL or FALSE, 
the user is asked to choose among four methods for selection based on time observation (most recent, date limit or case by case).}

\item{defaultParameters}{logical; should the default HeRAMS table parameters (e.g. column names) be used? If not, the user is able to modify it.}

\item{type}{logical; should the facilities be filtered based on their types.}

\item{ownership}{logical; should the facilities be filtered based on their ownership.}

\item{status}{logical; should the facilities be filtered based on their status.}

\item{building}{logical; should the facilities be filtered based on the building condition.}

\item{equipment}{logical; should the facilities be filtered based on the equipment condition.}

\item{functionality}{logical; should the facilities be filtered based on their functionality.}

\item{accessibility}{logical; should the facilities be filtered based on their accessibility.}

\item{support}{logical; should the facilities be filtered based on whether they receive support from partners.}

\item{services}{logical; should the facilities be filtered based on health services information.}

\item{partners}{logical; should the facilities also be filtered on the different possible supporting partners (ignored if support is FALSE)}

\item{barriers}{logical; should the facilities also be filtered on the causes of possible impairment (e.g. service not available).}
}
\description{
Filter the HeRAMS health facility raw table for different analysis scenario and export a table that contains only 
the selected facilities.
}
\details{
The selection is recorded within a text file (selected_hf.txt) stored in the scenario folder. Different
analysis scenario create new scenario folders. In the same scenario folder different 'raw' sub-folders may be created
depending on the original Excel document modification time, and the selection of observations based on time.
}
