% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeRAMS_impairment_values.R
\name{HeRAMS_impairment_values}
\alias{HeRAMS_impairment_values}
\title{HeRAMS impairment code values}
\usage{
HeRAMS_impairment_values()
}
\description{
Internal function to retrieve the values that indicate that there is an impairment
}
\keyword{internal}
