% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeRAMS_partnership_values.R
\name{HeRAMS_partnership_values}
\alias{HeRAMS_partnership_values}
\title{HeRAMS partnership code values}
\usage{
HeRAMS_partnership_values()
}
\description{
Internal function to retrieve the values that indicate that there is partner support
}
\keyword{internal}
