% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeRAMS_stop_filtering.R
\name{HeRAMS_stop_filtering}
\alias{HeRAMS_stop_filtering}
\title{HeRAMS stop filtering}
\usage{
HeRAMS_stop_filtering()
}
\description{
Internal function to retrieve the values that stop the questionnaire
}
