% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeRAMS_table_parameters.R
\name{HeRAMS_table_parameters}
\alias{HeRAMS_table_parameters}
\title{HeRAMS table parameters}
\usage{
HeRAMS_table_parameters()
}
\description{
Internal function to retrieve a list of the variable codes for filtering the HeRAMS table
}
\keyword{internal}
