% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeRAMS_table_subset.R
\name{HeRAMS_table_subset}
\alias{HeRAMS_table_subset}
\title{HeRAMS table subset}
\usage{
HeRAMS_table_subset(
  tibT,
  tibC,
  varCol,
  stopQuest = TRUE,
  codeName = NULL,
  stopLst = NULL,
  tempDir,
  barriers,
  codeColumns,
  impairmentValues,
  partners,
  partnershipValues
)
}
\arguments{
\item{tibT}{\code{tibble} object; HeRAMS table with labels.}

\item{tibC}{\code{tibble} object; HeRAMS table with codes.}

\item{varCol}{character; column name for filtering.}

\item{stopQuest}{logical; is it possible that the questionnaire is stopped at the specified column ?}

\item{codeName}{character; name of the column as it its indicated in the column code and stop lists.}

\item{stopLst}{list; list with the values for different attributes that stops the questionnaire.}

\item{tempDir}{character; temporary folder.}

\item{barriers}{logical; should the filtering process take into account the causes/barriers if there is an impairment (e.g., service partially available)}

\item{partnershipValues}{character; values that indicate that there is partner support. 
Has to be compatible with regular expression (regex).}

\item{impairValues}{character; values that indicate that there is an impairment (e.g., service partially available). 
Has to be compatible with regular expression (regex).}
}
\value{
a list of two \code{tibble} objects; The updated tibbles, the one with labels and the other one with codes.
}
\description{
Internal function used for subsetting the HeRAMS table and track record.
}
\keyword{internal}
