% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_best_cov.R
\name{hf_best_cov}
\alias{hf_best_cov}
\title{hf_best_cov}
\usage{
hf_best_cov(
  workDir,
  catchShp,
  popRaster,
  catchHfColName,
  nTot,
  adminCheck = FALSE,
  hfShp = NULL,
  hfHfColName = NULL,
  npAdmin = NULL,
  adminShp = NULL,
  adminColName = NULL
)
}
\arguments{
\item{workDir}{character; working directory that contains the inputs}

\item{catchShp}{character; name of the catchment shapefile WITHOUT extension}

\item{popRaster}{character; name of the population raster file WITH extension}

\item{catchHfColName}{character; name of the health facility name column in the catchment shapefile}

\item{nTot}{integer; number of health facilities to be selected}

\item{adminCheck}{logical; whether a minimum of health facilities per administrative unit must be selected. FALSE by default}

\item{hfShp}{character; name of the health facility shapefile WITHOUT extension}

\item{hfHfColName}{character; name of the health facility name column in the health facility shapefile}

\item{npAdmin}{integer; minimum number of health facilities to be selected per administrative unit (required if adminCheck = TRUE)}

\item{adminShp}{character; name of the administrative unit shapefile WITHOUT extention (required if adminCheck = TRUE)}

\item{adminColName}{character; name of the administrative unit name column in the administrative unit shapefile (required if adminCheck = TRUE)}
}
\description{
Select the health facilities that offer the best population coverage using their 
catchments and a raster of population. This function also offers the possibility to ensure that N 
facilities are selected for each administrative unit.
}
\details{
The outputs are a table with the selected facilities with their names, the population they cover, and the cumulative sum of the covered population. 
A bar plot representing the cumulative sum of the covered population for the selected facilities is also generated. The algorithm works recursively. 
It orders the facility based and the total population located within their catchments, it selects the first facility and removes the population covered 
in its catchment in case we would have catchments that overlap, to avoid counting the same population multiple times. 
It repeats the process until the number of facilities to be selected is reached.
}
