% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ts_fast.R
\name{multi_ts_fast}
\alias{multi_ts_fast}
\title{Multiple travel scenarios (alternative to multi_ts)}
\usage{
multi_ts_fast(inputFolder, adminLayerName, landcoverFile)
}
\arguments{
\item{inputFolder}{character; the path to the input folder, which must contains an administrative unit shapefile, a merged 
landcover imported from AccessMod and an Excel or CSV table for each travel scenario.}

\item{adminLayerName}{character; the name of the administrative unit layer (without extension)}

\item{landcoverFile}{character; the file name of the original merged landcover (with extension)}
}
\description{
Function that allows to handle different travel scenarios for different administrative units. It is faster than multi_ts and
recommended when the outputs of this latter can not be well processed in AccessMod. It creates an updated 
merged landcover and an updated travel scenario table (potentially much smaller that the one created by multi_ts). 
As for multi_ts, the required inputs are an administrative shapefile, a merged landcover
imported from AccessMod, and an Excel (or CSV) table for each travel scenario. The function requires a folder that only
contains the required inputs. When running the function, the user is asked to select the attribute table column of 
the shapefile that refers to each administrative unit (with no special character), and then to select the travel scenario for each administrative unit.
When modifying a scenario, the user must re-run the function and should not modify the outputs directly.
}
\details{
An output folder called out is created within the input folder, as well as a subfolder whose name 
is based on the system time that contains three outputs (e.g. ./out/20220826104842):
\itemize{
\item The updated landcover raster
\item The updated travel scenario table
\item A table that relates the different administrative units to the different travel scenarios
}
The function works are the following:
\itemize{
\item Check of the tables that are in the input folder: missing data, column names, values available for all landcover classes, values for the "mode" column, only one value per class, and speed in numerical format and positive or equal to zero.
\item Console printing of the shapefile attribute table, and selection of the column used to determine the administrative unit. 
\item Interactive selection of the scenario (based on the table names) for each administrative unit, and creation of a table that relates the units and the scenarios.
\item Loop over each scenario
\itemize{
\item Landcover raster clip based on the administrative units that are related to the scenario i.
\item Check if the scenario i has entries that are identical to entries from already processed scenarios
\item Creation of new classes if necessary
\item Reclassification of the raster depending on the two previous point
\item Update of the final travel scenario table
}
\item Merging of the rasters
\item Writing the final raster, the final table, and the table that relates the different administrative units and the different travel scenarios..
}
}
