% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_inputs.R
\name{process_inputs}
\alias{process_inputs}
\title{Process Input Layers}
\usage{
process_inputs(
  mainPath,
  region,
  all = FALSE,
  mostRecent = FALSE,
  alwaysProcess = FALSE,
  defaultMethods = FALSE,
  changeRes = NULL,
  newRes = NULL,
  popCorrection = NULL,
  gridRes = NULL
)
}
\arguments{
\item{mainPath}{character; the parent directory of the country/region name folder}

\item{region}{character; the country name}

\item{all}{logical; should all the available inputs be processed? If FALSE, the user is interactively asked to select the available
inputs to be processed.}

\item{mostRecent}{logical; should the most recent input be selected? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on date and time.}

\item{alwaysProcess}{logical; should always the input be processed? If alwaysProcess = FALSE and if the input has already
been processed, the user is interactively asked whether they want to process it or not.}

\item{defaultMethods}{logical; should be the default methods be used for projecting and resampling, respectively. These
are the 'bilinear' method for projecting and the 'sum' or the 'bilinear' for the resampling, depending on if the new resolution
is lower or higher.}

\item{changeRes}{logical; does the user want to change the raster resolution? If NULL, the resolution is printed and it is
interactively asked the user if they want to change it. IF FALSE, there is no resampling.}

\item{newRes}{numeric; new resolution in meters. Ignored if the newRes is FALSE. If NULL and if \code{changeRes} is TRUE,
the user is interactively asked to provide the new resolution.}

\item{popCorrection}{logical; should the raster correction algorithm be run. If it is NULL, the user is interactively asked
whether they want to run it or not.}

\item{gridRes}{numeric; the resolution of the grid shapefile used for correcting the raster. Ignored if popCorrection is FALSE.
If NULL and popCorrection is TRUE, the user is interactively asked to provide the grid resolution.}
}
\description{
Process any input layer and copy it to its corresponding folder
}
\details{
A 'processed' boundary shapefile is required for processing any other inputs. A 'processed' population raster is required
for processing any other raster. These conditions are taken into account and the processing of these
layers is performed even if they are not selected, if 'processed' layers are not available.
}
