% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pop.R
\name{process_pop}
\alias{process_pop}
\title{Process Population Raster}
\usage{
process_pop(
  mainPath,
  region,
  border,
  epsg,
  mostRecent,
  defaultMethods,
  changeRes,
  newRes,
  popCorrection,
  gridRes
)
}
\arguments{
\item{mainPath}{character; the parent directory of the country/region name folder}

\item{region}{character; the country name}

\item{border}{\code{sf} object; a boundary shapefile}

\item{epsg}{character; string that can be used as input in \code{raster::crs()} to describe a projection and datum}

\item{mostRecent}{logical; should the most recent input be selected? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on file creation time.}

\item{defaultMethods}{logical; should be the default methods be used for projecting and resampling, respectively. These
are the 'bilinear' method for projecting and the 'sum' or the 'bilinear' for the resampling, depending on if the new resolution
is lower or higher.}

\item{changeRes}{logical; does the user want to change the raster resolution? If NULL, the resolution is printed and it is
interactively asked the user if they want to change it. IF FALSE, there is no resampling.}

\item{newRes}{numeric; new resolution in meters. Ignored if the newRes is FALSE. If NULL and if \code{changeRes} is TRUE,
the user is interactively asked to provide the new resolution.}

\item{popCorrection}{logical; should the raster correction algorithm be run. If it is NULL, the user is interactively asked
whether they want to run it or not.}

\item{gridRes}{numeric; the resolution of the grid shapefile used for correcting the raster. Ignored if popCorrection is FALSE.
If NULL and popCorrection is TRUE, the user is interactively asked to provide the grid resolution.}
}
\value{
a list of length 2; The first element is the processed \code{SpatRaster} object and the second element is the selected
projection method (for track record)
}
\description{
Internal function used to process the population raster and copy it to its corresponding process folder
}
\details{
The algorithm for correcting the population raster works as following: it creates a grid shapefile, it sums up the population
in each of the cell considering  both the 'raw' and the 'processed' population raster. Then a ratio is calculated between both
values and related to each grid cell. The grid shapefile is rasterized using the ratios as values, and finally
the 'processed' raster is multiplied by the rasterized ratio. The lower is the grid resolution, the finer is the correction.
}
