% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raster.R
\name{process_raster}
\alias{process_raster}
\title{Process Raster Layer}
\usage{
process_raster(ras, border, epsg, projMeth)
}
\arguments{
\item{ras}{\code{SpatRaster} object; a raw input raster}

\item{border}{\code{sf} object; a boundary shapefile}

\item{epsg}{character; string that can be used as input in \code{raster::crs()} to describe a projection and datum}

\item{projMeth}{character; method used for estimating the new cell values. If NULL, the user is interactively
asked to select one of the available methods for \code{terra::project} function.}
}
\value{
a list of length 2; The first element is the processed \code{SpatRaster} object and the second element is the selected
projection method (for track record)
}
\description{
Internal function used to crop, mask and project rasters
}
\keyword{internal}
