% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_shapefile.R
\name{process_shapefile}
\alias{process_shapefile}
\title{Project and Clip Shapefile}
\usage{
process_shapefile(shp, epsg, inputName)
}
\arguments{
\item{shp}{\code{sf} object to be processed}

\item{epsg}{character; string that can be used as input in \code{sf::st_crs()} to describe a projection and datum}

\item{inputName}{character; the input name for console printing}
}
\value{
\code{sf} object
}
\description{
Internal function used to project and clip shapefiles
}
