% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_raster.R
\name{resample_raster}
\alias{resample_raster}
\title{Resample Raster Layer}
\usage{
resample_raster(ras1, ras0, rasInit, resampMeth)
}
\arguments{
\item{ras1}{\code{SpatRaster} object to be resampled}

\item{ras0}{\code{SpatRaster} object with the geometry that \code{ras1} should be resampled to}

\item{rasInit}{\code{SpatRaster} object; original 'raw' raster (before projecting)}

\item{resampMeth}{character; method used for estimating the new cell values. If NULL, the user is interactively
asked to select one of the available methods for \code{terra::resample} function.}
}
\value{
a list of length 2; the first element is the processed \code{SpatRaster} object and the second element is the selected
resampling method (for track record).
}
\description{
Internal function used to resample rasters
}
