% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioJSON.R
\name{scenarioJSON}
\alias{scenarioJSON}
\title{Create JSON files for multiple travel scenarios}
\usage{
scenarioJSON(inputFolder, landcoverFile, outputFolder, fileName)
}
\arguments{
\item{inputFolder}{character; the path to the input folder, which must contains an administrative unit shapefile, a merged 
landcover imported from AccessMod and an Excel or CSV table for each travel scenario.}

\item{landcoverFile}{character; the file name of the original merged landcover (with extension).}

\item{outputFolder}{character; the output folder where the JSON file will be written.}

\item{fileName}{character; the name of the output JSON file (without extension).}
}
\description{
From multiple travel scenarios table, the function creates a JSON file that can be used when running an AccessMod replay by command line.
}
\details{
A JSON file is created in the output folder.
}
