% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_folder.R
\name{select_folder}
\alias{select_folder}
\title{Select folders}
\usage{
select_folder(x, msg)
}
\arguments{
\item{x}{character vector; folder names}

\item{msg}{character; message to be print for selection}
}
\value{
character vector
}
\description{
Internal function used to subset elements from a character vector that correspond to folder names
}
