% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_hf_classes.R
\name{select_hf_classes}
\alias{select_hf_classes}
\title{Health facility attribute class selection (HeRAMS data)}
\usage{
select_hf_classes(categories, instructions)
}
\arguments{
\item{categories}{character; available values}

\item{instructions}{character; message to be displayed for the selection process}
}
\description{
Internal function used for an interactive class selection for a specific attribute
}
