% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_input.R
\name{select_input}
\alias{select_input}
\title{Select Input}
\usage{
select_input(folders, msg, mostRecent = FALSE)
}
\arguments{
\item{folders}{character; input folder names, usually obtained with the \code{check_exists} function}

\item{msg}{character; used to modify the presentation of the different options}

\item{mostRecent}{logical; should the most recent input be selected? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on time.}
}
\value{
character; time of the input folder creation
}
\description{
Internal function that is used to select an input when multiple ones are available
}
