% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_scenario.R
\name{select_scenario}
\alias{select_scenario}
\title{Select HeRAMS Scenario (internal function)}
\usage{
select_scenario(scenarioDirs)
}
\arguments{
\item{scenarioDirs}{character; Scenarios' folders}
}
\description{
Select HeRAMS Scenario (internal function)
}
