% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_scenarioTime.R
\name{select_scenarioTime}
\alias{select_scenarioTime}
\title{Select HeRAMS Scenario Folder (internal function)}
\usage{
select_scenarioTime(scenario)
}
\arguments{
\item{scenario}{character; Scenario folder}
}
\description{
Select HeRAMS Scenario Folder (internal function)
}
\keyword{internal}
