% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_HeRAMS_impairment_values.R
\name{set_HeRAMS_impairment_values}
\alias{set_HeRAMS_impairment_values}
\title{Set HeRAMS impairment values}
\usage{
set_HeRAMS_impairment_values(impairValues)
}
\arguments{
\item{stopLst}{list; internal data with default values}
}
\description{
Internal function to modify (if needed) the default values for the codes that indicate that there is an impairment
}
