% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_HeRAMS_key_values.R
\name{set_HeRAMS_key_values}
\alias{set_HeRAMS_key_values}
\title{Set HeRAMS key values}
\usage{
set_HeRAMS_key_values(values, msg)
}
\arguments{
\item{values}{character; default values (regex compatible)}
}
\description{
Internal function to modify (if needed) the default values for the key codes
}
