% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_HeRAMS_stop.R
\name{set_HeRAMS_stop}
\alias{set_HeRAMS_stop}
\title{Set HeRAMS stop responses}
\usage{
set_HeRAMS_stop(stopLst)
}
\arguments{
\item{stopLst}{list; internal data with default values}
}
\description{
Internal function to modify (if needed) the default values for the responses that stop the questionnaire
}
\keyword{internal}
