% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_projection.R
\name{set_projection}
\alias{set_projection}
\title{Set Projection}
\usage{
set_projection(
  mainPath,
  location,
  mostRecent = FALSE,
  alwaysSet = FALSE,
  bestCRS = FALSE,
  testMode = FALSE
)
}
\arguments{
\item{mainPath}{character; the parent directory of the location folder}

\item{location}{character; the location folder name}

\item{mostRecent}{logical; should the most recent downloaded boundary shapefile be selected? If FALSE and if there 
are multiple available inputs, the user is interactively asked to select the input based on download time.}

\item{alwaysSet}{logical; should the projected coordinate reference system always be set, even if it has already been 
set? If FALSE and if the projected coordinate reference system has already been set the user is 
interactively asked whether they want to set it again or not.}

\item{bestCRS}{logical; should the projected coordinate reference system be set automatically based on the "best-fit" 
projected coordinate reference system? If FALSE, the user is interactively asked to select the projected coordinate reference 
system from a list of suitable reference systems.}

\item{testMode}{logical; used for testing. If TRUE labels of processed inputs are not interactively asked.}
}
\description{
Set the projected coordinate reference system of the project based on the extent of the boundary shapefile.
}
\details{
The "best-fit" and the suitable projected coordinate reference systems are obtained with the 
\code{suggest_top_crs} and the \code{suggest_crs}, respectively, from the \pkg{crsuggest} package.
Both function work by analyzing the extent of the spatial dataset and comparing it to the area extents
in the EPSG's coordinate reference system database.
}
\examples{
# Replace workDir with the actual path to your working directory
\dontrun{
mainPath <- "workDir"
initiate_project(mainPath)}

# Replace myLocation with the location name you are working on (workDir subfolder)
\dontrun{
location <- "myLocation"
download_boundaries(mainPath, location, adminLevel = 1, type = "gbOpen", alwaysDownload = TRUE)
set_projection(mainPath, location, mostRecent = TRUE, alwaysSet = TRUE, bestCRS = TRUE)}
}
