% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_quietly.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stop Quietly}
\usage{
stop_quietly(msg)
}
\arguments{
\item{msg}{Character string}
}
\description{
Internal function that allows to stop a function with a non-error message
}
\examples{
stop_quietly("This is a non-error message");
}
\keyword{internal}
