% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_subset.R
\name{tibble_subset}
\alias{tibble_subset}
\title{Tibble Subset And Track Record}
\usage{
tibble_subset(tib, var, tempDir)
}
\arguments{
\item{tib}{\code{tibble} object;}

\item{var}{character vector; variable names (from filter_hf function)}

\item{tempDir}{character; temporary folder}
}
\value{
A \code{tibble} object
}
\description{
Internal function used in the \code{filter_hf} function
}
