% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_urban_areas.R
\docType{data}
\name{world_urban_areas}
\alias{world_urban_areas}
\title{This layer represents the major urban areas of the world as shaded polygons (ESRI)}
\format{
Polygon shapefile
}
\source{
https://hub.arcgis.com/datasets/schools-BE::world-urban-areas/about
\describe{
  This layer represents many major urban areas of the world as shaded polygons and is intended for cartographic display rather than analysis.
  Not all major cities throughout the world are represented.
}
}
\usage{
world_urban_areas
}
\description{
Data updated on November 11, 2018 and published on June 4, 2020. This work is licensed under the Esri Master License Agreement.
}
\keyword{datasets}
