% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_input.R
\name{copy_input}
\alias{copy_input}
\title{Copy Input}
\usage{
copy_input(mainPath, country, input)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{country}{character; the country folder name}

\item{input}{character; the absolute path of the raw input. Can be a single path (corresponding to e.g. a health facility table) 
or a vector of different paths (corresponding to e.g. shapefile documents). Can also be the path of a folder. In this case, the function
will copy all the files that are inside this folder.}
}
\description{
Copy a manually downloaded input to the project corresponding directory before being processed. To avoid possible further conflicts, 
the input has to be a "raw" input (i.e. not projected).
}
\details{
The user is interactively asked to select the 'input' folder destination.
}
\examples{
# Replace workDir with the actual path to your working directory
\dontrun{
mainPath <- "workDir"
initiate_project(mainPath, country)}

# Replace myCountry with the country name you are working on (workDir subfolder)
# Replace myInputPath with the actual path to the file to be copied or to the directory
# that contains the files to be copied (e.g., shapfile documents)
\dontrun{
country <- "myCountry"
inputPath <- "myInputPath"
copy_input(mainPath, country, inputPath)}
}
