% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hf_shapefile.R
\name{create_hf_shapefile}
\alias{create_hf_shapefile}
\title{Create a health facility shapefile}
\usage{
create_hf_shapefile(
  mainPath,
  region,
  mostRecentBoundaries = TRUE,
  mostRecentTable = FALSE,
  lonlat = TRUE,
  epsg = NULL,
  alwaysProcess = FALSE
)
}
\arguments{
\item{mainPath}{character; the parent directory of the country/region name folder}

\item{region}{character; the country name}

\item{mostRecentTable}{logical; should the most recent processed health facility table be used? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on date and time.}

\item{lonlat}{logical; are the coordinates indicated in the health facility table given in lon/lat?}

\item{epsg}{numeric or character (coerced to character); ESPG code - Coordinate systems worldwide (EPSG/ESRI).
Required only when lonlat = FALSE}

\item{alwaysProcess}{logical; should always a new shapefile be created? If alwaysProcess = FALSE and if a shapefile has already
been created from the same input table, the user is aksed whether they want to create a new shapefile or not.}

\item{mostRecentBoudaries}{logical; should the most recent processed boundary shapefile be used? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on date and time.}
}
\description{
Create a point shapefile of health facilities based on a pre-processed HeRAMS health facility table obtained with the
filter_hf function.
}
