% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dem.R
\name{download_dem}
\alias{download_dem}
\title{Download Digital Elevation Model}
\usage{
download_dem(mainPath, region, alwaysDownload = FALSE, mostRecent = FALSE)
}
\arguments{
\item{mainPath}{character; the parent directory of the country/region name folder}

\item{region}{character; the country/region name}

\item{alwaysDownload}{logical; should the raster always be downloaded, even if it has already been 
downloaded? If FALSE and if the raster has already been downloaded the user is 
interactively asked whether they want to download it again or not.}

\item{mostRecent}{logical; should the most recent boundary shapefile be selected to define the required DEM tiles? 
If FALSE and if there are multiple available inputs, the user is interactively asked to select the input based on date 
and time.}
}
\description{
Download a SRTM 90m resolution (3 arc-seconds) DEM for the entire country and copy it to its corresponding folder.
}
\details{
The function first download a SRTM tile grid shapefile from \url{https://github.com/sikli/srtm_country}.
The SRTM tiles to be downloaded are selected based on the extent of the boundary shapefile and are downloaded using the 
\code{getData} function from the \pkg{raster} package. If there are multiple tiles, it produces a mosaic.
}
