% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_landcover.R
\name{download_landcover}
\alias{download_landcover}
\title{Download Landcover}
\usage{
download_landcover(
  mainPath,
  location,
  globalLandCover = NULL,
  pathStoreLandCover = NULL,
  mostRecent = FALSE
)
}
\arguments{
\item{mainPath}{character; the parent directory of the location folder}

\item{location}{character; the location folder name}

\item{globalLandCover}{character; the path to the global land cover data (Cpernicus 100m discrete classification);
if NULL, the function will download the data from Zenodo and copy it to pathStoreLandCover}

\item{pathStoreLandCover}{character; the path to the folder where the global land cover data should be stored. When it is NULL
it saves the data in a temporary directory.}

\item{mostRecent}{logical; should the most recent boundary shapefile be selected to define the required landcover area?
If FALSE and if there are multiple available inputs, the user is interactively asked to select the input based on file creation
time.}
}
\description{
Download the Land Cover 100 m from the Copernicus Global Land Service and copy it to its corresponding folder.
}
\details{
The function will take the global land cover data and crop it to the extent of the boundaries of the location.
}
\examples{
# Replace workDir with the actual path to your working directory
\dontrun{
mainPath <- "workDir"
initiate_project(mainPath)}

# Replace myLocation with the location name you are working on (workDir subfolder)
\dontrun{
location <- "myLocation"
download_boundaries(mainPath, location, adminLevel = 1, type = "gbOpen", alwaysDownload = TRUE)
download_landcover(mainPath, location, globalLandCover = NULL, mostRecent = TRUE)}
}
