% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osm.R
\name{download_osm}
\alias{download_osm}
\title{Download Open Street Map Layers}
\usage{
download_osm(
  x,
  mainPath,
  country,
  alwaysDownload = FALSE,
  countryName = TRUE,
  mostRecent = NULL
)
}
\arguments{
\item{x}{character; target layer. Can be 'roads', 'waterLines' or 'naturalPolygons'}

\item{mainPath}{character; the parent directory of the country folder}

\item{country}{character; the country folder name}

\item{alwaysDownload}{logical; should the shapefile always be downloaded, even if it has already been 
downloaded? If FALSE and if the shapefile has already been downloaded the user is 
interactively asked whether they want to download it again or not.}

\item{countryName}{logical; should the country name be used to match with the \code{osm.pbf} file? If FALSE, it is the extent
of the boundary shapefile that is matched with the \code{osm.pbf} file.}

\item{mostRecent}{logical; should the most recent boundary shapefile be selected? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on file creation time.}
}
\description{
Download Open Street map shapefiles corresponding to 'roads', 'rivers' or any other natural feature and copy them
to their corresponding folders.
}
