% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_population.R
\name{download_population}
\alias{download_population}
\title{Download population raster}
\usage{
download_population(mainPath, region, alwaysDownload = FALSE)
}
\arguments{
\item{mainPath}{character; the parent directory of the country/region name folder}

\item{region}{character; the country/region name}

\item{alwaysDownload}{logical; should the raster always be downloaded, even they have already been 
downloaded? If FALSE and if the raster has already been downloaded the user is 
interactively asked whether they want to download it again or not.}
}
\description{
Download a population raster from the World Pop FTP
}
\details{
The function accesses the World Pop FTP \url{ftp://ftp.worldpop.org.uk/GIS/Population/} and uses an internal
function (\code{navigate_ftp}) in order to interactively navigate through the folders and select the population raster.
}
