% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_population.R
\name{download_population}
\alias{download_population}
\title{Download Population Raster}
\usage{
download_population(
  mainPath,
  country,
  alwaysDownload = FALSE,
  testMode = FALSE
)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{country}{character; the country folder name}

\item{alwaysDownload}{logical; should the raster always be downloaded, even if it has already been 
downloaded? If FALSE and if the raster has already been downloaded the user is 
interactively asked whether they want to download it again or not.}

\item{testMode}{logical; to be ignored. Only used when testing the function with testthat.}
}
\description{
Download a population raster from the World Pop FTP and copy it to its corresponding folder.
}
\details{
The function accesses the World Pop FTP and uses an internal function (\code{navigate_ftp}) in order to interactively navigate 
through the folders and select the population raster to be downloaded. The ISO code retrieved internally by the \code{get_param} function 
is used to match the country FTP folder when available.
}
\examples{
# Replace workDir with the actual path to your working directory
\dontrun{
mainPath <- "workDir"
initiate_project(mainPath)}

# Replace myCountry with the country name you are working on (workDir subfolder)
\dontrun{
country <- "myCountry"
download_population(mainPath, country, alwaysDownload = TRUE)}
}
