% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hf.R
\name{filter_hf}
\alias{filter_hf}
\title{Filter health facilities}
\usage{
filter_hf(
  mainPath,
  country,
  pathTable,
  scenario = NULL,
  mostRecentObs = NULL,
  barriers = TRUE
)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{country}{character; the country folder name}

\item{pathTable}{character; path to the HeRAMS Excel Table}

\item{scenario}{character; a string of three characters that correspond to the scenario folder suffix like '001', '002'...'010'...'099'...'100'
The criteria for selection are those of the specified scenario. If NULL, an interactive selection by attribute is run in the console.}

\item{mostRecentObs}{logical; should the most recent observation per health facility be taken into account? If NULL or FALSE, 
the user is asked to choose among four methods for selection based on time observation (most recent, date limit and most recent, closest to a specific date or case by case).}
}
\description{
Filter the HeRAMS health facility raw table based on a set of variables and export a table that contains only 
the selected facilities.
}
\details{
The selection is recorded within a text file (selected_hf.txt) stored in the scenario folder. Different
selection criteria create new scenario folders. In the same scenario folder different 'raw' sub-folders may be created
depending on the original Excel document modification time, and the selection of observations based on time.
}
