% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_param}
\alias{get_param}
\title{Get Project Parameters}
\usage{
get_param(mainPath, region, param)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{region}{character; the country folder name}

\item{param}{character; 'COUNTRY' to get the original country name, 'ISO' to get the ISO alpha-3 country code or 'EPSG'
to get the EPSG code for layer projection}
}
\description{
Internal function that is used to access the config.txt of the project and get a specific parameter (e.g. ISo alpha-3 code).
}
