% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_project.R
\name{initiate_project}
\alias{initiate_project}
\title{Initiate Project}
\usage{
initiate_project(
  mainPath,
  allowInteractivity = TRUE,
  city = FALSE,
  name = NULL,
  testMode = FALSE
)
}
\arguments{
\item{mainPath}{character; a path where the country folder will be created}

\item{allowInteractivity}{logical; whether to enable interactivity. \code{TRUE} by default.}

\item{city}{logical; whether to focus on cities instead of countries. \code{FALSE} by default.}

\item{name}{character; country or city name when \code{allowInteractivity} is set to \code{FALSE}. Must match perfectly either one of the names included
in inAccessMod::country_list (country) or inAccessMod::city_list (city). For cities, can also be the name of the city combined with the country ISO2 code
when > 1 city have the same name (ex. 'Vancouver'). In this case the format is the following: city name, white space, hyphen, space, code (ex. 'Zurich - CH').
This parameter is ignored when \code{allowInteractivity} is set to \code{TRUE}.}

\item{testMode}{logical; \code{FALSE} by default. Can be ignored. used for testing the function in the testthat context.}
}
\description{
Select the country, get the ISO 3166-1 alpha-3 country code, store it in a config.txt file and create the directory 
main structure for the project. This function also creates a log.txt file that will record and track the main operations 
related to the project.
}
\details{
The final structure arises when downloading and processing the data with the corresponding functions,
and it allows multiple 'raw' inputs and multiple 'processed' outputs for each input. This can be useful when 
performing different analyses for the same country (e.g. when we have updated data).
}
\examples{
# Replace workDir with the actual path to your working directory
\dontrun{
mainPath <- "workDir"
initiate_project(mainPath)}
}
