% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_landcover.R
\name{label_landcover}
\alias{label_landcover}
\title{Label landcover layer}
\usage{
label_landcover(
  mainPath,
  country,
  mostRecent,
  overwrite = FALSE,
  defaultLabels = TRUE
)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{country}{character; the country folder name}

\item{mostRecent}{logical; should the most recent input be selected? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on file creation time.}

\item{overwrite}{logical; in case it exists, should the output be overwritten?}

\item{defaultLabels}{logical; should the Copernicus default labels be associated with the the landcover raster values.
If FALSE, the user is interactively asked to enter the label names for each value.}
}
\description{
Creates a CSV table with landcover codes and labels that can be imported into AccessMod.
}
\details{
The function writes a CSV table with two columns (classes and labels) into the same folder as the processed landcover layer.
}
