% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_categories.R
\name{select_categories}
\alias{select_categories}
\title{Select by attribute}
\usage{
select_categories(sfObject, columnName)
}
\arguments{
\item{sfObject}{\code{sf} object}

\item{columnName}{character; the column name corresponding to the attribute used for filtering the shapefile}
}
\value{
A list of length 2; the first element is a \code{sf} object and the second element is the selected attribute values
}
\description{
Internal function used to select the OSM shapefiles by attribute
}
