% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_herams_table_parameters.R
\name{set_HeRAMS_table_parameters}
\alias{set_HeRAMS_table_parameters}
\title{Set HeRAMS table parameters}
\usage{
set_HeRAMS_table_parameters(tableParam, regex = TRUE)
}
\arguments{
\item{tableParam}{list; internal data with default values}
}
\description{
Internal function to modify (if needed) the default values for column codes
}
\keyword{internal}
