% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_regions.R
\name{subset_regions}
\alias{subset_regions}
\title{Subset Regions}
\usage{
subset_regions(mainPath, country, mostRecent = FALSE)
}
\arguments{
\item{mainPath}{character; the parent directory of the country folder}

\item{country}{character; the country folder name}

\item{mostRecent}{logical; should the most recent boundary shapefile be selected? If FALSE and if there are multiple
available inputs, the user is interactively asked to select the input based on file creation time.}
}
\description{
Function that allows to clip specific country regions from the 'raw' boundary shapefile when requiring analysis at the sub-national level
}
\examples{
# Replace workDir with the actual path to your working directory
\dontrun{
mainPath <- "workDir"
initiate_project(mainPath)}

# Replace myCountry with the country name you are working on (workDir subfolder)
\dontrun{
country <- "myCountry"
download_boundaries(mainPath, country, adminLevel = 1, type = "gbOpen", alwaysDownload = TRUE)
subset_regions(mainPath, country, mostRecent = TRUE)}
}
